setup.py remove unusable babel command
Since setuptools has upgraded version, a check of issubclass of
`Command` is applied on cmdclass dictionnary. With this check it is
impossible to run command with:

        python setup.py compile_catalog

Because `babel_wrapper.compile_catalog` is a method. Anyways, this
check is not done when the build execute `compile_catalog`, so:

        python setup.py install

still works.

To resolve the problem of others commands, we only need to execute
command via pybabel