classful: remove form.validate_on_submit()
validate_on_submit() is an alias for:
is_submitted(): if there is an active request (POST, PUT, PATCH, DELETE)
validate(): validate the form

Since we have separate view for each type of request, we don't need
is_submitted()
And we don't want a double validation (here and confd).
Sometime, we fill form with dynamic choices, ex:
choices for members of switchboard is all wazo users. So if we
validate the form, we will need to refetch all users twice (on
GET/INDEX to fill an empty form with choices and on PUT/POST to
validate the choices ...)