mallow: return missing if no input data
form.data will return an empty string if a value is given or not given.
But we need to make the difference, ex:

When creating a conference, we cannot fill the field 'music_on_hold'.
Suppose the default value in confd for music_on_hold is 'default'.
Without the following modification: the key 'music_on_hold' will be
sent with the value None. But it's not what we want. We just want to
not sent key music_on_hold