add possibility to nest destination
ex:
{'type': 'hangup',
 'cause': 'busy',
 'timeout': 5}

When we chose the value for 'type', we want to show the next key. In
this case it's the key 'cause'. We show the key timeout only after
having chosen the value for 'cause'

So the 'cause' key can use the same code of the 'type' key to show
the next field