int tests: set window size to 1920x1080
reason: When the Display size is not 1920x1080, there are a strange behavior
with the webdriver. e.g.:  browser is cropped or others similar bugs.
Maybe is due to my default resolution, but anyways that seems to cause
some bugs with tests:

selenium.common.exceptions.MoveTargetOutOfBoundsException: Message:
(825.875, 726.9000244140625) is out of bounds of viewport width (1024)
and height (692)

heigh is seem cropped to 692 in this case, but is set to 768 ...